Moved exp to expressions and removed all unused structs
