+++ src/lib.rs
@@ -155,7 +155,7 @@ pub use bin::{Bin, Bins};
 pub use client::Client;
 pub use commands::particle_type::ParticleType;
 pub use errors::{Error, ErrorKind, Result};
-pub use exp::regex_flag::RegexFlag;
+pub use expressions::regex_flag::RegexFlag;
 pub use key::Key;
 pub use net::Host;
 pub use operations::{MapPolicy, MapReturnType, MapWriteMode};
@@ -183,7 +183,7 @@ mod batch;
 mod client;
 mod cluster;
 mod commands;
-pub mod exp;
+pub mod expressions;
 mod msgpack;
 mod net;
 pub mod operations;
