+++ tests/src/exp.rs
@@ -15,7 +15,7 @@
 use crate::common;
 use env_logger;
 
-use aerospike::exp::{ExpType, Expression, FilterExpression};
+use aerospike::expressions::*;
 use aerospike::ParticleType;
 use aerospike::*;
 use std::sync::Arc;
@@ -53,9 +53,9 @@ fn expression_compare() {
 
     // EQ
     let rs = test_filter(
-        Expression::eq(
-            Expression::int_bin("bin".to_string()),
-            Expression::int_val(1),
+        eq(
+            int_bin("bin".to_string()),
+            int_val(1),
         ),
         &set_name,
     );
@@ -64,9 +64,9 @@ fn expression_compare() {
 
     // NE
     let rs = test_filter(
-        Expression::ne(
-            Expression::int_bin("bin".to_string()),
-            Expression::int_val(1),
+        ne(
+            int_bin("bin".to_string()),
+            int_val(1),
         ),
         &set_name,
     );
@@ -75,9 +75,9 @@ fn expression_compare() {
 
     // LT
     let rs = test_filter(
-        Expression::lt(
-            Expression::int_bin("bin".to_string()),
-            Expression::int_val(10),
+        lt(
+            int_bin("bin".to_string()),
+            int_val(10),
         ),
         &set_name,
     );
@@ -86,9 +86,9 @@ fn expression_compare() {
 
     // LE
     let rs = test_filter(
-        Expression::le(
-            Expression::int_bin("bin".to_string()),
-            Expression::int_val(10),
+        le(
+            int_bin("bin".to_string()),
+            int_val(10),
         ),
         &set_name,
     );
@@ -97,9 +97,9 @@ fn expression_compare() {
 
     // GT
     let rs = test_filter(
-        Expression::gt(
-            Expression::int_bin("bin".to_string()),
-            Expression::int_val(1),
+        gt(
+            int_bin("bin".to_string()),
+            int_val(1),
         ),
         &set_name,
     );
@@ -108,9 +108,9 @@ fn expression_compare() {
 
     // GE
     let rs = test_filter(
-        Expression::ge(
-            Expression::int_bin("bin".to_string()),
-            Expression::int_val(1),
+        ge(
+            int_bin("bin".to_string()),
+            int_val(1),
         ),
         &set_name,
     );
@@ -126,14 +126,14 @@ fn expression_condition() {
 
     // AND
     let rs = test_filter(
-        Expression::and(vec![
-            Expression::eq(
-                Expression::int_bin("bin".to_string()),
-                Expression::int_val(1),
+        and(vec![
+            eq(
+                int_bin("bin".to_string()),
+                int_val(1),
             ),
-            Expression::eq(
-                Expression::string_bin("bin2".to_string()),
-                Expression::string_val("1".to_string()),
+            eq(
+                string_bin("bin2".to_string()),
+                string_val("1".to_string()),
             ),
         ]),
         &set_name,
@@ -143,14 +143,14 @@ fn expression_condition() {
 
     // OR
     let rs = test_filter(
-        Expression::or(vec![
-            Expression::eq(
-                Expression::int_bin("bin".to_string()),
-                Expression::int_val(1),
+        or(vec![
+            eq(
+                int_bin("bin".to_string()),
+                int_val(1),
             ),
-            Expression::eq(
-                Expression::int_bin("bin".to_string()),
-                Expression::int_val(3),
+            eq(
+                int_bin("bin".to_string()),
+                int_val(3),
             ),
         ]),
         &set_name,
@@ -160,9 +160,9 @@ fn expression_condition() {
 
     // NOT
     let rs = test_filter(
-        Expression::not(Expression::eq(
-            Expression::int_bin("bin".to_string()),
-            Expression::int_val(1),
+        not(eq(
+            int_bin("bin".to_string()),
+            int_val(1),
         )),
         &set_name,
     );
@@ -178,9 +178,9 @@ fn expression_data_types() {
 
     // INT
     let rs = test_filter(
-        Expression::eq(
-            Expression::int_bin("bin".to_string()),
-            Expression::int_val(1),
+        eq(
+            int_bin("bin".to_string()),
+            int_val(1),
         ),
         &set_name,
     );
@@ -189,9 +189,9 @@ fn expression_data_types() {
 
     // STRING
     let rs = test_filter(
-        Expression::eq(
-            Expression::string_bin("bin2".to_string()),
-            Expression::string_val("1".to_string()),
+        eq(
+            string_bin("bin2".to_string()),
+            string_val("1".to_string()),
         ),
         &set_name,
     );
@@ -199,9 +199,9 @@ fn expression_data_types() {
     assert_eq!(count, 1, "STRING Test Failed");
 
     let rs = test_filter(
-        Expression::eq(
-            Expression::float_bin("bin3".to_string()),
-            Expression::float_val(2f64),
+        eq(
+            float_bin("bin3".to_string()),
+            float_val(2f64),
         ),
         &set_name,
     );
@@ -209,9 +209,9 @@ fn expression_data_types() {
     assert_eq!(count, 1, "FLOAT Test Failed");
 
     let rs = test_filter(
-        Expression::eq(
-            Expression::blob_bin("bin4".to_string()),
-            Expression::blob_val(format!("{}{}", "blob", 5).into_bytes()),
+        eq(
+            blob_bin("bin4".to_string()),
+            blob_val(format!("{}{}", "blob", 5).into_bytes()),
         ),
         &set_name,
     );
@@ -219,9 +219,9 @@ fn expression_data_types() {
     assert_eq!(count, 1, "BLOB Test Failed");
 
     let rs = test_filter(
-        Expression::ne(
-            Expression::bin_type("bin".to_string()),
-            Expression::int_val(ParticleType::NULL as i64),
+        ne(
+            bin_type("bin".to_string()),
+            int_val(ParticleType::NULL as i64),
         ),
         &set_name,
     );
@@ -237,21 +237,21 @@ fn expression_rec_ops() {
 
     // dev size 0 because in-memory
     let rs = test_filter(
-        Expression::le(Expression::device_size(), Expression::int_val(0)),
+        le(device_size(), int_val(0)),
         &set_name,
     );
     let count = count_results(rs);
     assert_eq!(count, 100, "DEVICE SIZE Test Failed");
 
     let rs = test_filter(
-        Expression::gt(Expression::last_update(), Expression::int_val(15000)),
+        gt(last_update(), int_val(15000)),
         &set_name,
     );
     let count = count_results(rs);
     assert_eq!(count, 100, "LAST UPDATE Test Failed");
 
     let rs = test_filter(
-        Expression::gt(Expression::since_update(), Expression::int_val(150)),
+        gt(since_update(), int_val(150)),
         &set_name,
     );
     let count = count_results(rs);
@@ -259,69 +259,69 @@ fn expression_rec_ops() {
 
     // Records dont expire
     let rs = test_filter(
-        Expression::le(Expression::void_time(), Expression::int_val(0)),
+        le(void_time(), int_val(0)),
         &set_name,
     );
     let count = count_results(rs);
     assert_eq!(count, 100, "VOID TIME Test Failed");
 
     let rs = test_filter(
-        Expression::le(Expression::ttl(), Expression::int_val(0)),
+        le(ttl(), int_val(0)),
         &set_name,
     );
     let count = count_results(rs);
     assert_eq!(count, 100, "TTL Test Failed");
 
-    let rs = test_filter(Expression::not(Expression::is_tombstone()), &set_name);
+    let rs = test_filter(not(is_tombstone()), &set_name);
     let count = count_results(rs);
     assert_eq!(count, 100, "TOMBSTONE Test Failed");
 
     let rs = test_filter(
-        Expression::eq(
-            Expression::set_name(),
-            Expression::string_val(set_name.clone()),
+        eq(
+            expressions::set_name(),
+            string_val(set_name.clone()),
         ),
         &set_name,
     );
     let count = count_results(rs);
     assert_eq!(count, 100, "SET NAME Test Failed");
 
-    let rs = test_filter(Expression::bin_exists("bin4".to_string()), &set_name);
+    let rs = test_filter(bin_exists("bin4".to_string()), &set_name);
     let count = count_results(rs);
     assert_eq!(count, 100, "BIN EXISTS Test Failed");
 
     let rs = test_filter(
-        Expression::eq(Expression::digest_modulo(3), Expression::int_val(1)),
+        eq(digest_modulo(3), int_val(1)),
         &set_name,
     );
     let count = count_results(rs);
     assert_eq!(count > 0 && count < 100, true, "DIGEST MODULO Test Failed");
 
     let rs = test_filter(
-        Expression::eq(Expression::key(ExpType::INT), Expression::int_val(50)),
+        eq(key(ExpType::INT), int_val(50)),
         &set_name,
     );
     let count = count_results(rs);
     // 0 because key is not saved
     assert_eq!(count, 0, "KEY Test Failed");
 
-    let rs = test_filter(Expression::key_exists(), &set_name);
+    let rs = test_filter(key_exists(), &set_name);
     let count = count_results(rs);
     // 0 because key is not saved
     assert_eq!(count, 0, "KEY EXISTS Test Failed");
 
     let rs = test_filter(
-        Expression::eq(Expression::nil(), Expression::nil()),
+        eq(nil(), nil()),
         &set_name,
     );
     let count = count_results(rs);
     assert_eq!(count, 100, "NIL Test Failed");
 
     let rs = test_filter(
-        Expression::regex_compare(
+        regex_compare(
             "[1-5]".to_string(),
             RegexFlag::ICASE as i64,
-            Expression::string_bin("bin2".to_string()),
+            string_bin("bin2".to_string()),
         ),
         &set_name,
     );
@@ -353,120 +353,120 @@ fn expression_commands() {
 
     // DELETE
     let key = as_key!(namespace, &set_name, 15);
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(16),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(16),
     ));
     let test = client.delete(&wpolicy, &key);
     assert_eq!(test.is_err(), true, "DELETE EXP Err Test Failed");
 
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(15),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(15),
     ));
     let test = client.delete(&wpolicy, &key);
     assert_eq!(test.is_ok(), true, "DELETE EXP Ok Test Failed");
 
     // PUT
     let key = as_key!(namespace, &set_name, 25);
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(15),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(15),
     ));
     let test = client.put(&wpolicy, &key, &[as_bin!("bin", 26)]);
     assert_eq!(test.is_err(), true, "PUT Err Test Failed");
 
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(25),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(25),
     ));
     let test = client.put(&wpolicy, &key, &[as_bin!("bin", 26)]);
     assert_eq!(test.is_ok(), true, "PUT Ok Test Failed");
 
     // GET
     let key = as_key!(namespace, &set_name, 35);
-    rpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(15),
+    rpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(15),
     ));
     let test = client.get(&rpolicy, &key, Bins::All);
     assert_eq!(test.is_err(), true, "GET Err Test Failed");
 
-    rpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(35),
+    rpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(35),
     ));
     let test = client.get(&rpolicy, &key, Bins::All);
     assert_eq!(test.is_ok(), true, "GET Ok Test Failed");
 
     // EXISTS
     let key = as_key!(namespace, &set_name, 45);
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(15),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(15),
     ));
     let test = client.exists(&wpolicy, &key);
     assert_eq!(test.is_err(), true, "EXISTS Err Test Failed");
 
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(45),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(45),
     ));
     let test = client.exists(&wpolicy, &key);
     assert_eq!(test.is_ok(), true, "EXISTS Ok Test Failed");
 
     // APPEND
     let key = as_key!(namespace, &set_name, 55);
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(15),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(15),
     ));
     let test = client.add(&wpolicy, &key, &[as_bin!("test55", "test")]);
     assert_eq!(test.is_err(), true, "APPEND Err Test Failed");
 
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(55),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(55),
     ));
     let test = client.add(&wpolicy, &key, &[as_bin!("test55", "test")]);
     assert_eq!(test.is_ok(), true, "APPEND Ok Test Failed");
 
     // PREPEND
     let key = as_key!(namespace, &set_name, 55);
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(15),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(15),
     ));
     let test = client.prepend(&wpolicy, &key, &[as_bin!("test55", "test")]);
     assert_eq!(test.is_err(), true, "PREPEND Err Test Failed");
 
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(55),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(55),
     ));
     let test = client.prepend(&wpolicy, &key, &[as_bin!("test55", "test")]);
     assert_eq!(test.is_ok(), true, "PREPEND Ok Test Failed");
 
     // TOUCH
     let key = as_key!(namespace, &set_name, 65);
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(15),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(15),
     ));
     let test = client.touch(&wpolicy, &key);
     assert_eq!(test.is_err(), true, "TOUCH Err Test Failed");
 
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(65),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(65),
     ));
     let test = client.touch(&wpolicy, &key);
     assert_eq!(test.is_ok(), true, "TOUCH Ok Test Failed");
 
     // SCAN
-    spolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(75),
+    spolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(75),
     ));
     match client.scan(&spolicy, namespace, &set_name, Bins::All) {
         Ok(records) => {
@@ -487,17 +487,17 @@ fn expression_commands() {
     let ops = vec![operations::add(&bin)];
 
     let key = as_key!(namespace, &set_name, 85);
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(15),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(15),
     ));
     let op = client.operate(&wpolicy, &key, &ops);
     assert_eq!(op.is_err(), true, "OPERATE Err Test Failed");
 
     let key = as_key!(namespace, &set_name, 85);
-    wpolicy.filter_expression = Some(Expression::eq(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(85),
+    wpolicy.filter_expression = Some(eq(
+        int_bin("bin".to_string()),
+        int_val(85),
     ));
     let op = client.operate(&wpolicy, &key, &ops);
     assert_eq!(op.is_ok(), true, "OPERATE Ok Test Failed");
@@ -509,9 +509,9 @@ fn expression_commands() {
         let key = as_key!(namespace, &set_name, i);
         batch_reads.push(BatchRead::new(key, &b));
     }
-    bpolicy.filter_expression = Some(Expression::gt(
-        Expression::int_bin("bin".to_string()),
-        Expression::int_val(84),
+    bpolicy.filter_expression = Some(gt(
+        int_bin("bin".to_string()),
+        int_val(84),
     ));
     match client.batch_get(&bpolicy, batch_reads) {
         Ok(results) => {
